adding more backend pois tests adding ownership permissions for pois adding register page

+++ backend/apps/pois/views.py
@@ -5,15 +5,11 @@ from rest_framework.permissions import IsAuthenticated
 from django.shortcuts import get_object_or_404
 from .models import POI
 from .serializers import POISerializer
+from .permissions import IsOwner
 
 
 class POIsView(APIView):
     permission_classes = (IsAuthenticated,)
-
-    def get(self, request):
-        pois = POI.objects.all()
-        serializer = POISerializer(pois, many=True)
-        return Response(serializer.data, status=status.HTTP_200_OK)
 
     def post(self, request):
         serializer = POISerializer(data=request.data)
@@ -22,17 +18,24 @@ class POIsView(APIView):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
+    def get(self, request):
+        pois = POI.objects.filter(created_by=request.user)
+        serializer = POISerializer(pois, many=True)
+        return Response(serializer.data, status=status.HTTP_200_OK)
+
 
 class POIView(APIView):
-    permission_classes = (IsAuthenticated,)
+    permission_classes = (IsAuthenticated, IsOwner,)
 
     def get(self, request, pk):
         poi = get_object_or_404(POI, pk=pk)
+        self.check_object_permissions(request, poi)
         serializer = POISerializer(poi)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def put(self, request, pk):
         poi = get_object_or_404(POI, pk=pk)
+        self.check_object_permissions(request, poi)
         serializer = POISerializer(poi, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -41,5 +44,6 @@ class POIView(APIView):
 
     def delete(self, request, pk):
         poi = get_object_or_404(POI, pk=pk)
+        self.check_object_permissions(request, poi)
         poi.delete()
         return Response({'msg': 'POI deleted successfully.'}, status=status.HTTP_204_NO_CONTENT)
